agregado controlador de inicio de sesión y middleware para la extracción de tokens; se implementó verificación de token en las rutas de blogs

+++ part4/blogList_backend/controllers/blog.controller.js
@@ -1,15 +1,36 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog.model')
 const User = require('../models/user.model')
+const jwt = require('jsonwebtoken')
 
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
+blogRouter.get('/:id', async (request, response) => {
+  try {
+    const blog = await Blog.findById(request.params.id).populate('user', {
+      username: 1,
+      name: 1,
+    })
+    if (blog) {
+      response.json(blog)
+    } else {
+      response.status(404).end()
+    }
+  } catch (error) {
+    response.status(400).json({ error: error.message })
+  }
+})
+
 blogRouter.post('/', async (request, response) => {
   const { title, author, url, likes, userId } = request.body
-  const user = await User.findById(userId)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   if (!title || !url) {
     return response.status(400).json({ error: 'title and url are required' })
@@ -55,8 +76,26 @@ blogRouter.put('/:id', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+      return response.status(404).json({ error: 'blog not found' })
+    }
+
+    if (blog.user.toString() !== decodedToken.id) {
+      return response.status(401).json({ error: 'unauthorized' })
+    }
+
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } catch (error) {
+    response.status(400).json({ error: error.message })
+  }
 })
 
 module.exports = blogRouter
